refactor(web): extract integration connection picker
Pull the MemberConnectionPicker out of agent-integrations-block into a
reusable integration-connect/integration-connection-picker.tsx, backed by
shared connectable-auth-keys + package-files helpers. Collapse the
run-overrides-panel per-authKey select loop to the flat
{ integrationId: connectionId } shape. No behaviour change.

Touches: agent-integrations-block, run-overrides-panel, diff-tab,
package-editor, unified-package-detail, locales (en/fr).

Verified standalone on feat/integrations: web tsc clean.
