+++ apps/web/src/components/package-detail/agent-integrations-block.tsx
@@ -1,55 +1,23 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import { useState } from "react";
-import { useQueryClient } from "@tanstack/react-query";
 import { useTranslation } from "react-i18next";
-import {
-  AlertTriangle,
-  Loader2,
-  Puzzle,
-  Users,
-  Check,
-  Plus,
-  Lock,
-  ChevronDown,
-  RefreshCw,
-  Settings,
-} from "lucide-react";
-import { useNavigate } from "react-router-dom";
-import { Button } from "@/components/ui/button";
-import { Badge } from "@/components/ui/badge";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+import { Loader2, Puzzle } from "lucide-react";
 import {
   useIntegrations,
   useIntegrationDetail,
   useIntegrationConnections,
   useAgentsConsumingIntegration,
-  useIntegrationAgentResolution,
   type AgentIntegrationEntry,
-  type IntegrationAgentResolution,
   type IntegrationAuthStatus,
-  type IntegrationCandidate,
   type IntegrationConnection,
   type IntegrationManifestView,
 } from "../../hooks/use-integrations";
-import {
-  useUpsertMemberIntegrationPin,
-  useDeleteMemberIntegrationPin,
-} from "../../hooks/use-member-integration-pins";
 import { InlineConnectButton } from "../integration-connect/inline-connect-button";
-import { FieldsConnectModal } from "../integration-connect/fields-connect-modal";
-import { useIntegrationOAuthPopup } from "../integration-connect/use-integration-oauth-popup";
 import { connectionDisplayLabel } from "../integration-connect/connection-label";
 import { pickDefaultAuth } from "../integration-connect/pick-default-auth";
+import { connectableAuthKeys } from "../integration-connect/connectable-auth-keys";
+import { IntegrationConnectionPicker } from "../integration-connect/integration-connection-picker";
 import { requiredScopesForAgent } from "@appstrate/core/integration";
-import { api } from "../../api";
 
 interface AgentIntegrationsBlockProps {
   entries: AgentIntegrationEntry[];
@@ -183,7 +151,7 @@ function IntegrationConnectionCard({
   if (showMemberPicker) {
     return (
       <CardShell title={displayName} subtitle={packageId}>
-        <MemberConnectionPicker
+        <IntegrationConnectionPicker
           integrationId={packageId}
           agentPackageId={agentPackageId!}
           manifest={detail.manifest}
@@ -206,29 +174,6 @@ function IntegrationConnectionCard({
       agentScopes={agentScopes}
     />
   );
-}
-
-/**
- * Auth keys the actor can actually start a connect flow on. oauth2 auths
- * require an admin-registered OAuth client (`hasOAuthClient`); without one
- * the server refuses connect with 403. Non-oauth2 auths (api_key / basic /
- * custom) carry no client and are always connectable via the fields modal.
- * Mirrors the gate the integration detail page applies.
- */
-function connectableAuthKeys(
-  manifest: IntegrationManifestView,
-  authStatuses: IntegrationAuthStatus[],
-): Set<string> {
-  const out = new Set<string>();
-  const declared = manifest.auths ?? {};
-  for (const key of Object.keys(declared)) {
-    if (declared[key]?.type !== "oauth2") {
-      out.add(key);
-      continue;
-    }
-    if (authStatuses.find((s) => s.auth_key === key)?.has_oauth_client) out.add(key);
-  }
-  return out;
 }
 
 /**
@@ -363,434 +308,6 @@ function buildReuseInfo(
     return t("detail.integrationReuseSingle", { account });
   }
   return t("detail.integrationReuseShared", { account, count: agentCount });
-}
-
-/**
- * Per-integration connection picker for the member, rendered as a rich
- * dropdown. Lists every accessible connection (own + shared-with-org)
- * with its name, auth type (OAuth / API key …), and who created it, plus
- * an "auto" entry (defer to the resolver) and "add a connection" entries
- * (one per declared auth) that launch the connect flow inline.
- *
- * Picks are persisted as `integration_pins` rows with `user_id` set —
- * the resolver sees them at cascade layer 4 on every run. Admin pins
- * (`user_id IS NULL`) at layer 1 lock the choice for everyone: when an
- * admin pin exists for this (agent, integration), the dropdown is
- * disabled and shows the pinned connection with a lock badge.
- */
-function MemberConnectionPicker({
-  integrationId,
-  agentPackageId,
-  manifest,
-  authStatuses,
-  displayName,
-  agentTools,
-  agentScopes,
-}: {
-  integrationId: string;
-  agentPackageId: string;
-  manifest: IntegrationManifestView;
-  authStatuses: IntegrationAuthStatus[];
-  displayName: string;
-  agentTools: string[] | "*" | undefined;
-  agentScopes: string[] | undefined;
-}) {
-  const { t } = useTranslation(["agents", "settings"]);
-  const { data: resolution, isPending } = useIntegrationAgentResolution(
-    integrationId,
-    agentPackageId,
-  );
-  const upsertPin = useUpsertMemberIntegrationPin();
-  const deletePin = useDeleteMemberIntegrationPin();
-  const { openPopup, isPending: oauthPending } = useIntegrationOAuthPopup();
-  const qc = useQueryClient();
-  const navigate = useNavigate();
-  const [fieldsAuthKey, setFieldsAuthKey] = useState<string | null>(null);
-
-  const auths = manifest.auths ?? {};
-  // Only auths the actor can actually connect: oauth2 needs an admin OAuth
-  // client (else the connect 403s); api_key/basic/custom always can. Without
-  // this the "add connection" entries offered a flow doomed to 403.
-  const connectable = connectableAuthKeys(manifest, authStatuses);
-  const authKeys = Object.keys(auths).filter((k) => connectable.has(k));
-  const typeLabel = (authKey: string): string | null => {
-    const type = auths[authKey]?.type;
-    return type ? t(`settings:integration.auth.type.${type}`) : null;
-  };
-  const fieldsAuth = fieldsAuthKey ? auths[fieldsAuthKey] : null;
-  // The whole verdict (cascade + scope diff) is computed server-side; a pin
-  // write or scope upgrade invalidates it so the dropdown re-resolves.
-  const refresh = () => qc.invalidateQueries({ queryKey: ["integrations"] });
-
-  if (isPending || !resolution) {
-    return (
-      <div data-testid={`member-picker-${integrationId}`}>
-        <Button variant="outline" size="sm" disabled className="h-7 gap-1.5 text-xs">
-          <Loader2 className="size-3 animate-spin" />
-        </Button>
-      </div>
-    );
-  }
-
-  const {
-    candidates,
-    status,
-    resolved_connection_id: resolvedConnectionId,
-    resolved_missing_scopes: resolvedMissingScopes,
-    admin_pinned_connection_id: adminPinnedConnectionId,
-    member_pinned_connection_id: memberPinnedConnectionId,
-    org_default_connection_id: orgDefaultConnectionId,
-    org_default_enforced: orgDefaultEnforced,
-    can_add_connection: canAddConnection,
-  } = resolution;
-
-  const ownerLabel = (c: IntegrationCandidate): string =>
-    c.is_own
-      ? t("detail.integrationMemberPicker.byYou")
-      : (c.owner_name ?? t("detail.integrationMemberPicker.ownerUnknown"));
-
-  // Locked when an admin force applies and the member can never override:
-  // a per-agent admin pin OR an enforced org default. Either way we render
-  // the read-only lock instead of the editable dropdown.
-  const lockedConnectionId =
-    adminPinnedConnectionId ?? (orgDefaultEnforced ? orgDefaultConnectionId : null);
-  if (lockedConnectionId) {
-    const pinned = candidates.find((c) => c.id === lockedConnectionId);
-    const label = pinned ? connectionDisplayLabel(pinned) : lockedConnectionId;
-    return (
-      <div data-testid={`member-picker-${integrationId}`}>
-        <Button
-          variant="outline"
-          size="sm"
-          disabled
-          className="h-7 justify-start gap-1.5 text-xs"
-          data-testid={`member-pick-locked-${integrationId}`}
-        >
-          <Lock className="size-3" />
-          <span className="truncate">{label}</span>
-          <Badge variant="secondary" className="ml-1 text-[0.6rem]">
-            {t("detail.integrationMemberPicker.adminLocked")}
-          </Badge>
-        </Button>
-      </div>
-    );
-  }
-
-  const current = memberPinnedConnectionId;
-  const resolvedConn = candidates.find((c) => c.id === resolvedConnectionId) ?? null;
-  const hasCandidates = candidates.length > 0;
-  const underScoped = resolvedMissingScopes.length > 0;
-
-  // Select the connection the user just added: "add a connection" both
-  // creates AND selects it (they added it to use it), persisting a member pin
-  // so the resolver lands on it instead of dropping into a blocked
-  // "must_choose". Only when the member hasn't already pinned an explicit
-  // choice — an existing pin is respected, not overridden.
-  const selectConnection = async (connectionId: string) => {
-    await upsertPin.mutateAsync({ agentPackageId, integrationId, connectionId });
-    await refresh();
-  };
-
-  const triggerConnect = async (authKey: string, opts?: { connectionId?: string }) => {
-    const auth = auths[authKey];
-    if (!auth) return;
-    if (auth.type === "oauth2") {
-      // Snapshot the accessible set so we can identify the just-created
-      // connection afterwards — the OAuth popup can't return its id, and a
-      // cancelled popup adds nothing, leaving the prior resolution intact.
-      // On a renew (connectionId supplied) the callback UPDATEs in place
-      // and the snapshot diff is empty — we skip the post-popup pin step.
-      const before = new Set(candidates.map((c) => c.id));
-      const hadPin = !!memberPinnedConnectionId;
-      const isRenew = !!opts?.connectionId;
-      // Forward the agent's per-tool inferred scopes so consent asks for what
-      // THIS agent needs — not just the integration's manifest defaults (the
-      // integration detail page is the surface that connects at defaults).
-      const scopes = requiredScopesForAgent({ manifest, authKey, agentTools, agentScopes });
-      await openPopup({
-        packageId: integrationId,
-        authKey,
-        ...(scopes.length ? { scopes } : {}),
-        // Account picker is noise on a renew — the user is re-authorising the
-        // existing identity, not picking a new one. Force-pick stays on fresh
-        // connects so "Add another" actually offers a different account.
-        ...(isRenew ? {} : { forceAccountSelect: true }),
-        ...(opts?.connectionId ? { connectionId: opts.connectionId } : {}),
-      });
-      if (isRenew || hadPin) {
-        await refresh();
-        return;
-      }
-      const fresh = await api<IntegrationAgentResolution>(
-        `/integrations/${encodeURI(integrationId)}/agent-resolution/${encodeURI(agentPackageId)}`,
-      );
-      const added = fresh.candidates.find((c) => !before.has(c.id));
-      if (added) await selectConnection(added.id);
-      else await refresh();
-    } else {
-      // api_key / basic / custom — no row-update path via the credentials
-      // modal yet; we surface the connect modal. The modal's `connectionId`
-      // wiring is OAuth-only for now (matches the single-writer convergence
-      // in integration-connections.ts for non-OAuth auths).
-      setFieldsAuthKey(authKey);
-    }
-  };
-
-  const triggerLabel = resolvedConn
-    ? connectionDisplayLabel(resolvedConn)
-    : status === "must_choose"
-      ? t("detail.integrationMemberPicker.chooseLabel")
-      : status === "stale"
-        ? t("detail.integrationMemberPicker.reconfigureLabel")
-        : t("detail.integrationMemberPicker.connectLabel");
-  // Warning visuals when there's no usable resolved connection OR the
-  // resolved one is under-scoped (run would be blocked).
-  const triggerWarn = status === "must_choose" || status === "stale" || underScoped;
-  const TriggerIcon = triggerWarn ? AlertTriangle : resolvedConn ? Users : Plus;
-
-  // Blocked for this member AND nothing to pick → dead end. Show a
-  // disabled, explanatory button instead of an empty dropdown.
-  if (!canAddConnection && !hasCandidates) {
-    return (
-      <div data-testid={`member-picker-${integrationId}`}>
-        <Button
-          variant="outline"
-          size="sm"
-          disabled
-          className="h-7 justify-start gap-1.5 text-xs"
-          data-testid={`member-pick-blocked-${integrationId}`}
-        >
-          <Lock className="size-3" />
-          <span className="truncate">{t("detail.integrationMemberPicker.blockedByAdmin")}</span>
-        </Button>
-      </div>
-    );
-  }
-
-  // No existing connection AND no auth the actor can connect on (every
-  // oauth2 auth lacks an admin-registered OAuth client) → point at the
-  // admin setup instead of an empty dropdown that would only 403.
-  if (!hasCandidates && authKeys.length === 0) {
-    return (
-      <div data-testid={`member-picker-${integrationId}`}>
-        <span
-          className="text-muted-foreground text-xs"
-          data-testid={`member-pick-no-client-${integrationId}`}
-        >
-          {t("settings:integration.auth.noClientHint")}
-        </span>
-      </div>
-    );
-  }
-
-  return (
-    <div data-testid={`member-picker-${integrationId}`}>
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button
-            variant="outline"
-            size="sm"
-            className={`h-7 justify-start gap-1.5 text-xs ${triggerWarn ? "text-amber-600 dark:text-amber-400" : ""}`}
-            data-testid={`member-pick-${integrationId}`}
-          >
-            <TriggerIcon className="size-3" />
-            <span className="max-w-[14rem] truncate">{triggerLabel}</span>
-            {status === "auto" && (
-              <span className="text-muted-foreground/70">
-                {t("detail.integrationMemberPicker.defaultBadge")}
-              </span>
-            )}
-            <ChevronDown className="size-3 opacity-60" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="start" className="max-w-[20rem]">
-          <DropdownMenuLabel className="text-[0.7rem]">
-            {t("detail.integrationMemberPicker.title")}
-          </DropdownMenuLabel>
-          {candidates.map((c) => {
-            const tl = typeLabel(c.auth_key);
-            const isDefault = !current && resolvedConnectionId === c.id;
-            // Only the connection owner can renew via OAuth — a foreign
-            // shared connection's tokens belong to someone else. We still
-            // let the actor pin a foreign needs_reconnection row (their
-            // pick survives once the owner renews it).
-            const canRenew =
-              c.needs_reconnection && c.is_own && auths[c.auth_key]?.type === "oauth2";
-            return (
-              <DropdownMenuItem
-                key={c.id}
-                onSelect={() =>
-                  upsertPin.mutate(
-                    { agentPackageId, integrationId, connectionId: c.id },
-                    { onSuccess: refresh },
-                  )
-                }
-                data-testid={`member-pick-option-${c.id}`}
-              >
-                <Check className={`size-3.5 ${resolvedConnectionId === c.id ? "" : "opacity-0"}`} />
-                <div className="flex min-w-0 flex-1 flex-col">
-                  <div className="flex items-center gap-1.5">
-                    <span className="truncate font-medium">{connectionDisplayLabel(c)}</span>
-                    {tl && (
-                      <Badge variant="outline" className="text-[0.6rem]">
-                        {tl}
-                      </Badge>
-                    )}
-                    {c.shared_with_org && (
-                      <Badge variant="secondary" className="text-[0.6rem]">
-                        {t("detail.integrationMemberPicker.sharedBadge")}
-                      </Badge>
-                    )}
-                    {c.missing_scopes.length > 0 && (
-                      <Badge variant="destructive" className="text-[0.6rem]">
-                        {t("detail.integrationMemberPicker.missingScopesBadge")}
-                      </Badge>
-                    )}
-                    {isDefault && (
-                      <span className="text-muted-foreground/70 text-[0.6rem]">
-                        {t("detail.integrationMemberPicker.defaultBadge")}
-                      </span>
-                    )}
-                  </div>
-                  <span className="text-muted-foreground truncate text-[0.65rem]">
-                    {t("detail.integrationMemberPicker.connectedBy", { owner: ownerLabel(c) })}
-                    {c.needs_reconnection &&
-                      ` · ${t("detail.integrationMemberPicker.needsReconnection")}`}
-                  </span>
-                </div>
-                {canRenew && (
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    className="ml-1 h-6 gap-1 px-2 text-[0.65rem] text-amber-600 hover:text-amber-700 dark:text-amber-400"
-                    disabled={oauthPending}
-                    onClick={(e) => {
-                      // Block the DropdownMenuItem's onSelect so the renew
-                      // click doesn't also pin the dead row.
-                      e.preventDefault();
-                      e.stopPropagation();
-                      void triggerConnect(c.auth_key, { connectionId: c.id });
-                    }}
-                    data-testid={`member-pick-renew-${c.id}`}
-                    aria-label={t("detail.integrationMemberPicker.renew")}
-                  >
-                    <RefreshCw className="size-3" />
-                    {t("detail.integrationMemberPicker.renew")}
-                  </Button>
-                )}
-              </DropdownMenuItem>
-            );
-          })}
-          {current && (
-            <DropdownMenuItem
-              onSelect={() =>
-                deletePin.mutate({ agentPackageId, integrationId }, { onSuccess: refresh })
-              }
-              data-testid={`member-pick-reset-${integrationId}`}
-            >
-              <Check className="size-3.5 opacity-0" />
-              <span className="text-muted-foreground">
-                {t("detail.integrationMemberPicker.resetToAuto")}
-              </span>
-            </DropdownMenuItem>
-          )}
-          {canAddConnection && hasCandidates && authKeys.length > 0 && <DropdownMenuSeparator />}
-          {canAddConnection &&
-            authKeys.map((k) => {
-              const tl = typeLabel(k);
-              return (
-                <DropdownMenuItem
-                  key={`add-${k}`}
-                  onSelect={() => void triggerConnect(k)}
-                  data-testid={`member-pick-add-${integrationId}-${k}`}
-                >
-                  <Plus className="size-3.5" />
-                  <span>
-                    {authKeys.length > 1 && tl
-                      ? t("detail.integrationMemberPicker.addVia", { label: tl })
-                      : t("detail.integrationMemberPicker.addConnection")}
-                  </span>
-                </DropdownMenuItem>
-              );
-            })}
-          <DropdownMenuSeparator />
-          {/* Escape hatch to the integration page for the full connection
-              management surface (rename, share-with-org, delete, OAuth client). */}
-          <DropdownMenuItem
-            onSelect={() => navigate(`/integrations/${integrationId}`)}
-            data-testid={`member-pick-manage-${integrationId}`}
-          >
-            <Settings className="size-3.5" />
-            <span>{t("detail.integrationMemberPicker.manageConnections")}</span>
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
-      {/* Resolved connection is under-scoped → the run is blocked
-          server-side (insufficient_scopes). Owner can upgrade via
-          incremental consent; a foreign owner can only be flagged. */}
-      {underScoped && resolvedConn && (
-        <div
-          className="mt-1.5 flex flex-col gap-1.5 rounded-md border border-amber-500/40 bg-amber-500/10 p-2 text-[0.7rem] text-amber-700 dark:text-amber-300"
-          data-testid={`member-pick-scope-warning-${integrationId}`}
-        >
-          <div className="flex items-center gap-1.5">
-            <AlertTriangle className="size-3 shrink-0" />
-            <span>
-              {resolution.resolved_owned_by_actor
-                ? t("detail.integrationMemberPicker.missingScopesOwn")
-                : t("detail.integrationMemberPicker.missingScopesForeign", {
-                    owner: ownerLabel(resolvedConn),
-                  })}
-            </span>
-          </div>
-          <span className="text-foreground/80 font-mono text-[0.65rem] break-words">
-            {resolvedMissingScopes.join(" ")}
-          </span>
-          {resolution.resolved_owned_by_actor &&
-            auths[resolvedConn.auth_key]?.type === "oauth2" && (
-              <div>
-                <Button
-                  size="sm"
-                  disabled={oauthPending}
-                  onClick={async () => {
-                    await openPopup({
-                      packageId: integrationId,
-                      authKey: resolvedConn.auth_key,
-                      scopes: resolvedMissingScopes,
-                      connectionId: resolvedConn.id,
-                    });
-                    // The OAuth callback updated the connection's granted
-                    // scopes server-side; refetch so the badge clears
-                    // instead of waiting for a window-focus refetch.
-                    await refresh();
-                  }}
-                  data-testid={`member-pick-upgrade-${integrationId}`}
-                >
-                  <RefreshCw className="mr-1 size-3" />
-                  {t("detail.integrationMemberPicker.upgradeButton")}
-                </Button>
-              </div>
-            )}
-        </div>
-      )}
-      {fieldsAuth && fieldsAuthKey && (
-        <FieldsConnectModal
-          open={true}
-          onClose={() => setFieldsAuthKey(null)}
-          packageId={integrationId}
-          authKey={fieldsAuthKey}
-          auth={fieldsAuth}
-          displayName={displayName}
-          // The fields connect returns the created connection, so select it
-          // directly — same "add = select it" rule as the OAuth path, and only
-          // when the member hasn't already pinned an explicit choice.
-          onConnected={(conn) => {
-            if (!memberPinnedConnectionId) void selectConnection(conn.id);
-          }}
-        />
-      )}
-    </div>
-  );
 }
 
 /**
