+++ apps/web/src/locales/en/agents.json
@@ -41,9 +41,6 @@
   "missingConnections.close": "Close",
   "missingConnections.retry": "Re-run",
   "missingConnections.resolved": "Connection ready — re-run",
-  "missingConnections.mustChoose.cta": "Manage connections",
-  "missingConnections.mustChoose.candidates_one": "{{count}} connection available — rename one, share, or disconnect the others to resolve the ambiguity:",
-  "missingConnections.mustChoose.candidates_other": "{{count}} connections available — rename one, share, or disconnect the others to resolve the ambiguity:",
   "missingConnections.mustChoose.pickPrompt_one": "Pick the connection to use for this run ({{count}} available):",
   "missingConnections.mustChoose.pickPrompt_other": "Pick the connection to use for this run ({{count}} available):",
   "missingConnections.mustChoose.retry": "Run with these picks",
@@ -53,6 +50,7 @@
   "detail.integrationMemberPicker.chooseLabel": "Choose one",
   "detail.integrationMemberPicker.reconfigureLabel": "Needs reconfiguration",
   "detail.integrationMemberPicker.resetToAuto": "Reset (automatic resolution)",
+  "detail.integrationMemberPicker.inherit": "Inherit (default resolution)",
   "detail.integrationMemberPicker.blockedByAdmin": "Connections blocked by admin",
   "detail.integrationMemberPicker.missingScopesBadge": "Missing permissions",
   "detail.integrationMemberPicker.missingScopesOwn": "This connection lacks some permissions the agent requires. Upgrade it to grant the missing ones.",
@@ -218,9 +216,6 @@
   "editor.integrationsEmpty": "No integrations imported. Visit the Integrations marketplace to import one.",
   "editor.integrationInactive": "Not active in this application",
   "editor.tabGeneral": "General",
-  "editor.tabContent.agent": "prompt.md",
-  "editor.tabContent.skill": "SKILL.md",
-  "editor.tabContent.mcp-server": "README.md",
   "editor.tabSchema": "Schemas",
   "editor.tabSkills": "Skills",
   "editor.tabRuntimeTools": "Runtime tools",
@@ -380,10 +375,6 @@
   "schedule.overridesHint": "These overrides apply on every execution of this schedule. Leave on \"Inherit\" to use the agent's default settings.",
   "schedule.connectionOverrides.label": "Integration connections",
   "schedule.connectionOverrides.hint": "Force a specific connection per integration the agent depends on. Leave on \"Inherit\" to use the default resolution (admin pin → user fallback) at each fire time.",
-  "schedule.connectionOverrides.inherit": "— Inherit (default resolution at runtime) —",
-  "schedule.connectionOverrides.noCandidate": "No connection available for this integration.",
-  "schedule.connectionOverrides.sharedSuffix": "{{label}} ⋯ shared",
-  "schedule.connectionOverrides.selectAria": "Connection override for {{authKey}}",
   "detail.configSectionProxy": "Proxy",
   "detail.configAlert": "Required configuration fields are not filled in.",
   "detail.tabApi": "API",
