+++ apps/web/src/locales/fr/agents.json
@@ -41,9 +41,6 @@
   "missingConnections.close": "Fermer",
   "missingConnections.retry": "Relancer le run",
   "missingConnections.resolved": "Connexion prête — relancez le run",
-  "missingConnections.mustChoose.cta": "Gérer les connexions",
-  "missingConnections.mustChoose.candidates_one": "{{count}} connexion disponible — renommez-en une, partagez ou déconnectez les autres pour lever l'ambiguïté :",
-  "missingConnections.mustChoose.candidates_other": "{{count}} connexions disponibles — renommez-en une, partagez ou déconnectez les autres pour lever l'ambiguïté :",
   "missingConnections.mustChoose.pickPrompt_one": "Choisissez la connexion à utiliser pour ce run ({{count}} disponible) :",
   "missingConnections.mustChoose.pickPrompt_other": "Choisissez la connexion à utiliser pour ce run ({{count}} disponibles) :",
   "missingConnections.mustChoose.retry": "Lancer avec ces choix",
@@ -53,6 +50,7 @@
   "detail.integrationMemberPicker.chooseLabel": "À choisir",
   "detail.integrationMemberPicker.reconfigureLabel": "À reconfigurer",
   "detail.integrationMemberPicker.resetToAuto": "Réinitialiser (résolution automatique)",
+  "detail.integrationMemberPicker.inherit": "Hériter (résolution par défaut)",
   "detail.integrationMemberPicker.blockedByAdmin": "Connexions bloquées par l'admin",
   "detail.integrationMemberPicker.missingScopesBadge": "Permissions manquantes",
   "detail.integrationMemberPicker.missingScopesOwn": "Cette connexion n'a pas toutes les permissions requises par l'agent. Mettez-la à niveau pour les autoriser.",
@@ -218,9 +216,6 @@
   "editor.integrationsEmpty": "Aucune intégration importée. Visitez le marketplace Intégrations pour en importer une.",
   "editor.integrationInactive": "Non active dans cette application",
   "editor.tabGeneral": "Général",
-  "editor.tabContent.agent": "prompt.md",
-  "editor.tabContent.skill": "SKILL.md",
-  "editor.tabContent.mcp-server": "README.md",
   "editor.tabSchema": "Schémas",
   "editor.tabSkills": "Skills",
   "editor.tabRuntimeTools": "Outils runtime",
@@ -380,10 +375,6 @@
   "schedule.overridesHint": "Ces paramètres s'appliquent à chaque exécution de cette planification. Laisser sur \"Hériter\" pour utiliser les paramètres par défaut de l'agent.",
   "schedule.connectionOverrides.label": "Connexions d'intégration",
   "schedule.connectionOverrides.hint": "Forcez une connexion spécifique pour chaque intégration dont dépend l'agent. Laisser sur « Hériter » pour utiliser la résolution par défaut (épingle admin → fallback utilisateur) au moment de chaque déclenchement.",
-  "schedule.connectionOverrides.inherit": "— Hériter (résolution par défaut au runtime) —",
-  "schedule.connectionOverrides.noCandidate": "Aucune connexion disponible pour cette intégration.",
-  "schedule.connectionOverrides.sharedSuffix": "{{label}} ⋯ partagé",
-  "schedule.connectionOverrides.selectAria": "Connexion forcée pour {{authKey}}",
   "detail.configSectionProxy": "Proxy",
   "detail.configAlert": "Des champs de configuration obligatoires ne sont pas remplis.",
   "detail.tabApi": "API",
